Seperating all the endpoints
